+++ src/test/java/org/corpus_tools/annis/gui/ExportPanelTest.java
@@ -2,12 +2,16 @@ package org.corpus_tools.annis.gui;
 
 import static com.github.mvysny.kaributesting.v8.LocatorJ._click;
 import static com.github.mvysny.kaributesting.v8.LocatorJ._get;
+import static com.github.mvysny.kaributesting.v8.LocatorJ._setValue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import com.github.mvysny.kaributesting.v8.MockVaadin;
 import com.google.common.collect.Sets;
 import com.vaadin.spring.internal.UIScopeImpl;
 import com.vaadin.ui.Button;
+import com.vaadin.ui.ComboBox;
+import org.corpus_tools.annis.gui.controlpanel.SearchOptionsPanel;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -62,7 +66,57 @@ class ExportPanelTest {
     // enabled
     _click(_get(panel, Button.class, spec -> spec.withCaption("Perform Export")));
     TestHelper.awaitCondition(30, downloadButton::isEnabled);
+  }
 
+  @SuppressWarnings({"rawtypes", "unchecked"})
+  @Test
+  void testTextColumnExportContextChange() throws Exception {
+    // Prepare query
+    ui.getQueryController().setQuery(
+        QueryGenerator.displayed().corpora(Sets.newHashSet("pcc2")).query("pos=\"NE\"").build());
+
+    // Click on the "More" button and then "Export"
+    PopupButton moreButton = _get(PopupButton.class, spec -> spec.withCaption("More"));
+    moreButton.setPopupVisible(true);
+    _click(_get(Button.class, spec -> spec.withCaption("Export")));
+
+    // Make sure the Export tab is there
+    ExportPanel panel = _get(ExportPanel.class);
+
+    // Change to text column exporter
+    ComboBox exporterSelection = _get(panel, ComboBox.class, spec -> spec.withCaption("Exporter"));
+    _setValue(exporterSelection, "TextColumnExporter");
+
+    // Get the search options panel that should be in sync
+    SearchOptionsPanel searchOptions = _get(SearchOptionsPanel.class);
+
+    // Check that the initial values are correct
+    ComboBox leftContextExportPanel =
+        _get(panel, ComboBox.class, spec -> spec.withCaption("Left Context"));
+    ComboBox rightContextExportPanel =
+        _get(panel, ComboBox.class, spec -> spec.withCaption("Right Context"));
+    assertEquals(5, leftContextExportPanel.getValue());
+    assertEquals(5, rightContextExportPanel.getValue());
+
+    ComboBox leftContextSearchOptions =
+        _get(searchOptions, ComboBox.class, spec -> spec.withCaption("Left Context"));
+    ComboBox rightContextSearchOptions =
+        _get(searchOptions, ComboBox.class, spec -> spec.withCaption("Right Context"));
+    assertEquals(5, leftContextSearchOptions.getValue());
+    assertEquals(5, rightContextSearchOptions.getValue());
+
+    assertEquals(5, ui.getQueryState().getLeftContext());
+    assertEquals(5, ui.getQueryState().getRightContext());
+
+    // Change the value of the combo boxes and test that the query state and the search panel
+    // options have changed
+    _setValue(leftContextExportPanel, 10);
+    _setValue(rightContextExportPanel, 20);
+
+    assertEquals(10, ui.getQueryState().getLeftContext());
+    assertEquals(20, ui.getQueryState().getRightContext());
+    assertEquals(10, leftContextSearchOptions.getValue());
+    assertEquals(20, rightContextSearchOptions.getValue());
   }
 
 }